CAS-10933 move stats framework files to new directory make symlinks to relevant .h files for backward compatibility for includes
