+++ scimath/CMakeLists.txt
@@ -13,7 +13,6 @@ Functionals/FuncExprData.cc
 Functionals/FunctionFactoryErrors.cc
 Functionals/FunctionalProxy.cc
 Functionals/SerialHelper.cc
-Mathematics/ClassicalStatisticsData.cc
 Mathematics/Combinatorics.cc
 Mathematics/FFTPack.cc
 Mathematics/FFTServer.cc
@@ -29,10 +28,11 @@ Mathematics/NumericTraits.cc
 Mathematics/RigidVector2.cc
 Mathematics/SCSL.cc
 Mathematics/SquareMatrix2.cc
-Mathematics/StatisticsData.cc
 Mathematics/VectorKernel.cc
 Mathematics/VanVleck.cc
-Mathematics/ZScoreCalculator.cc
+StatsFramework/ClassicalStatisticsData.cc
+StatsFramework/StatisticsData.cc
+StatsFramework/ZScoreCalculator.cc
 )
 
 if (FFTW3_FOUND)
@@ -231,13 +231,10 @@ Mathematics/AutoDiffRep.h
 Mathematics/AutoDiffRep.tcc
 Mathematics/AutoDiffX.h
 Mathematics/ChauvenetCriterionStatistics.h
-Mathematics/ChauvenetCriterionStatistics.tcc
 Mathematics/ClassicalStatistics.h
-Mathematics/ClassicalStatistics.tcc
 Mathematics/ClassicalStatisticsData.h
 Mathematics/Combinatorics.h
 Mathematics/ConstrainedRangeStatistics.h
-Mathematics/ConstrainedRangeStatistics.tcc
 Mathematics/ConvolveGridder.h
 Mathematics/ConvolveGridder.tcc
 Mathematics/Convolver.h
@@ -248,14 +245,12 @@ Mathematics/FFTPack.h
 Mathematics/FFTServer.h
 Mathematics/FFTW.h
 Mathematics/FitToHalfStatistics.h
-Mathematics/FitToHalfStatistics.tcc
 Mathematics/FitToHalfStatisticsData.h
 Mathematics/GaussianBeam.h
 Mathematics/Geometry.h
 Mathematics/Gridder.h
 Mathematics/Gridder.tcc
 Mathematics/HingesFencesStatistics.h
-Mathematics/HingesFencesStatistics.tcc
 Mathematics/HistAcc.h
 Mathematics/HistAcc.tcc
 Mathematics/Interpolate2D.h
@@ -293,23 +288,48 @@ Mathematics/SquareMatrix.tcc
 Mathematics/StatAcc.h
 Mathematics/StatAcc.tcc
 Mathematics/StatsDataProvider.h
-Mathematics/StatsDataProvider.tcc
 Mathematics/StatisticsAlgorithm.h
-Mathematics/StatisticsAlgorithm.tcc
 Mathematics/StatisticsAlgorithmFactory.h
-Mathematics/StatisticsAlgorithmFactory.tcc
 Mathematics/StatisticsAlgorithmFactoryData.h
 Mathematics/StatisticsData.h
 Mathematics/StatisticsIncrementer.h
-Mathematics/StatisticsIncrementer.tcc
 Mathematics/StatisticsTypes.h
-Mathematics/StatisticsTypes.tcc
 Mathematics/StatisticsUtilities.h
-Mathematics/StatisticsUtilities.tcc
 Mathematics/VanVleck.h
 Mathematics/VectorKernel.h
 Mathematics/ZScoreCalculator.h
 DESTINATION include/casacore/scimath/Mathematics
+)
+
+install (FILES
+StatsFramework/ChauvenetCriterionStatistics.h
+StatsFramework/ChauvenetCriterionStatistics.tcc
+StatsFramework/ClassicalStatistics.h
+StatsFramework/ClassicalStatistics.tcc
+StatsFramework/ClassicalStatisticsData.h
+StatsFramework/ConstrainedRangeStatistics.h
+StatsFramework/ConstrainedRangeStatistics.tcc
+StatsFramework/FitToHalfStatistics.h
+StatsFramework/FitToHalfStatistics.tcc
+StatsFramework/FitToHalfStatisticsData.h
+StatsFramework/HingesFencesStatistics.h
+StatsFramework/HingesFencesStatistics.tcc
+StatsFramework/StatsDataProvider.h
+StatsFramework/StatsDataProvider.tcc
+StatsFramework/StatisticsAlgorithm.h
+StatsFramework/StatisticsAlgorithm.tcc
+StatsFramework/StatisticsAlgorithmFactory.h
+StatsFramework/StatisticsAlgorithmFactory.tcc
+StatsFramework/StatisticsAlgorithmFactoryData.h
+StatsFramework/StatisticsData.h
+StatsFramework/StatisticsIncrementer.h
+StatsFramework/StatisticsIncrementer.tcc
+StatsFramework/StatisticsTypes.h
+StatsFramework/StatisticsTypes.tcc
+StatsFramework/StatisticsUtilities.h
+StatsFramework/StatisticsUtilities.tcc
+StatsFramework/ZScoreCalculator.h
+DESTINATION include/casacore/scimath/StatsFramework
 )
 
 install (FILES
@@ -322,3 +342,5 @@ DESTINATION include/casacore/scimath
 add_subdirectory (Fitting/test ${EXCL_ALL})
 add_subdirectory (Functionals/test ${EXCL_ALL})
 add_subdirectory (Mathematics/test ${EXCL_ALL})
+add_subdirectory (StatsFramework/test ${EXCL_ALL})
+
