+++ scimath/StatsFramework/HingesFencesStatistics.tcc
