+++ scimath/StatsFramework/StatisticsAlgorithm.tcc
