+++ scimath/StatsFramework/StatisticsAlgorithmFactory.tcc
