+++ scimath/StatsFramework/StatisticsAlgorithmFactoryData.h
@@ -1,0 +1,60 @@
+//# Copyright (C) 2000,2001
+//# Associated Universities, Inc. Washington DC, USA.
+//#
+//# This library is free software; you can redistribute it and/or modify it
+//# under the terms of the GNU Library General Public License as published by
+//# the Free Software Foundation; either version 2 of the License, or (at your
+//# option) any later version.
+//#
+//# This library is distributed in the hope that it will be useful, but WITHOUT
+//# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+//# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Library General Public
+//# License for more details.
+//#
+//# You should have received a copy of the GNU Library General Public License
+//# along with this library; if not, write to the Free Software Foundation,
+//# Inc., 675 Massachusetts Ave, Cambridge, MA 02139, USA.
+//#
+//# Correspondence concerning AIPS++ should be addressed as follows:
+//#        Internet email: aips2-request@nrao.edu.
+//#        Postal address: AIPS++ Project Office
+//#                        National Radio Astronomy Observatory
+//#                        520 Edgemont Road
+//#                        Charlottesville, VA 22903-2475 USA
+//#
+
+#ifndef SCIMATH_STATSALGORITHMFACTORYDATA_H
+#define SCIMATH_STATSALGORITHMFACTORYDATA_H
+
+#include <casacore/scimath/Mathematics/FitToHalfStatisticsData.h>
+
+namespace casacore {
+
+// define data structures used by tatisticsAlgorithmFactory
+
+class StatisticsAlgorithmFactoryData {
+
+public:
+
+    template <class AccumType> struct FitToHalfData {
+        FitToHalfStatisticsData::CENTER center;
+        // fit to half data portion to use
+        FitToHalfStatisticsData::USE_DATA side;
+        // fit to half center value (only relevent if center=CVALUE)
+        AccumType centerValue;
+    };
+
+    struct ChauvenetData {
+        Double zScore;
+        Int maxIter;
+    };
+
+    ~StatisticsAlgorithmFactoryData() {};
+
+private:
+    StatisticsAlgorithmFactoryData() {};
+};
+
+}
+
+#endif
