+++ scimath/StatsFramework/StatisticsIncrementer.tcc
