+++ scimath/StatsFramework/test/tStatisticsAlgorithmFactory.cc
